Add TRIM.IN, add counter-av dist, polish rat macro
The TRIM.IN files allow TRIM to run in batch mode, and store the
configuration of the air layer in a file.

Both the height from the origin, and the distance to the av of the
counter are important parameters (though they sum to the inner
distance), and so may as well print out both values in
compute-parameters.

Change generated particles to alphas (oops), lower event count for now,
and add tracking to event generation.